workout-curation/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
geoalchemy2==0.15.2
pgvector==0.3.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
apscheduler==3.10.4
sentence-transformers==3.3.1
torch==2.5.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
